[unreal]  Fix module search function to prioritize index.js over package.json, and presence of .js in module folder name (Fixes issue #1223) (修复模块导入问题)
This commit addresses two issues related to module import. Firstly, it fixes the issue where the module search function was not prioritizing index.js when both index.js and package.json existed in the same directory, which goes against JavaScript convention. Secondly, it addresses the issue where the folder name of a module contained ".js" (e.g. bn.js), causing the search function to fail. This fix ensures that the module search function searches for a JavaScript module in the specified directory and prioritizes index.js when both files exist in the same directory.